+++ tests/tests_io/test_questaal.py
@@ -1,10 +1,7 @@
 import os
 import unittest
 
-try:
-    from importlib.resources import files as ilr_files
-except ImportError:  # Python < 3.9
-    from importlib_resources import files as ilr_files
+from importlib.resources import files as ilr_files
 
 import numpy as np
 from pymatgen.core.lattice import Lattice