+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Change Log
 Unreleased
 ----------
 
+- Bump minimum supported Python version to 3.10, in line with upstream Python support. (https://devguide.python.org/versions/)
+
 - Bump minimum phonopy version to 2.7.0 (July 2020) to cleanly replace
   deprecated/removed PhonopyAtoms.get_number_of_atoms()